service/requirements.txt
Flask==1.1.1
requests==2.21.0
CloudEvent==1.2.0
Parse==1.19.0
